requirements.txt
aws-cdk-lib==2.160.0
constructs>=10.0.0,<11.0.0
cdklabs.generative-ai-cdk-constructs==0.1.270
aws-cdk.aws-lambda-python-alpha>=2.141.0a0
cdklabs.generative-ai-cdk-constructs>=0.1.158
requests>=2.32.3
aws-lambda-powertools>=3.0.0